refactor: add helper function dir and functions
