+++ src/commands/A11yTreeCommands.ts
@@ -2,6 +2,10 @@ import * as vscode from 'vscode';
 import * as fs from 'fs';
 import * as path from 'path';
 import * as puppeteer from 'puppeteer';
+
+import getTabIndex from '../helpers/getTabIndex';
+import checkNode from '../helpers/checkNode';
+
 import {
   User,
   A11yTreeCommands,
@@ -38,7 +42,6 @@ import {
 
 const a11yTreeCommands: A11yTreeCommands = {
   async handleFetchTree(
-    port: number,
     panel: vscode.WebviewPanel,
     context: vscode.ExtensionContext,
     url: string,
@@ -63,74 +66,14 @@ const a11yTreeCommands: A11yTreeCommands = {
       // visit passed in url
       await page.goto(url);
 
-      // search current tree of elements to find the node that is focused
-      async function findFocusedNode(
-        node: AccessibilityNode
-      ): Promise<AccessibilityNode | undefined> {
-        // return node that is focused (node that was tabbed to) and return it
-        if (node.focused) {
-          return node;
-        }
-        // if node has children, check children for focused node
-        for (const child of node.children || []) {
-          const focusedNode = await findFocusedNode(child);
-          if (focusedNode) {
-            return focusedNode;
-          }
-        }
-
-        return undefined;
-      }
-
       // set link count so we can track the first 3 focusable nodes for skip link
       let linkCount = 0;
       // set up partial compliance object to build out full compliance object later
       let compliance = { h1: null, skipLink: null };
 
       // tab through whole page, check for skip links and compile all tabbable nodes to an array
-      async function getTabIndex(
-        page: any,
-        compliance: Compliance
-      ): Promise<AccessibilityNode[]> {
-        let snap: AccessibilityNode | null =
-          // capture state of page
-          await page.accessibility.snapshot();
-        // setup array to collect all tabbable nodes/elements
-        const tabIndex: AccessibilityNode[] = [];
 
-        while (snap) {
-          await page.keyboard.press('Tab');
-          snap = await page.accessibility.snapshot();
-
-          if (!snap) {
-            break; // Ensure snap is not null before proceeding
-          }
-
-          const focusedNode = await findFocusedNode(snap);
-          if (!focusedNode) {
-            break;
-          }
-
-          if (
-            !compliance.skipLink &&
-            (focusedNode.role = 'link') &&
-            linkCount < 3
-          ) {
-            if (testForSkipLink(focusedNode)) {
-              compliance.skipLink = focusedNode;
-            }
-            linkCount++;
-          }
-          // janky, but need to pass node to add compliances;
-          checkNode(focusedNode, compliance);
-
-          tabIndex.push(focusedNode);
-        }
-
-        return tabIndex;
-      }
-
-      const tabIndex = await getTabIndex(page, compliance);
+      const tabIndex = await getTabIndex(page, compliance, linkCount);
 
       // Generate Puppeteer accessibility tree snapshot and cast it to our custom AccessibilityTree type
       const a11yTree =
@@ -159,7 +102,6 @@ const a11yTreeCommands: A11yTreeCommands = {
       };
 
       // const outputFolder = path.join(context.extensionPath, 'results');
-
       // const treeResultPath = path.join(outputFolder, 'a11y-tree.json');
 
       // Save results to local files for testing
@@ -248,6 +190,11 @@ export async function getUserSelectedProjectDirectoryName(): Promise<
   return selectedFolder;
 }
 
+// !!!!!!! lets remove this !!!!!!!!!!
+
+let lastHeadingLevel = 0;
+let firstH1 = false;
+
 // Adding compliance details to each element of the accessibility tree
 function guidelineCreator(
   tree: AccessibilityTree,
@@ -268,175 +215,6 @@ function treeCrawl(
       treeCrawl(child, compliance);
     }
   } else {
-    checkNode(node, compliance);
+    checkNode(node, compliance, lastHeadingLevel, firstH1);
   }
 }
-
-// Add guideline properties based on node roles, which are link, button, and heading
-function checkNode(node: AccessibilityNode, compliance: Compliance) {
-  node.compliance = true;
-  node.complianceDetails = '';
-  switch (node.role) {
-    case 'link':
-      if (testLink(node)) {
-        node.compliance = false;
-        node.complianceDetails += ' link text must provide meaningful context';
-      }
-      break;
-    case 'button':
-      node.compliance = node.name ? true : false;
-      node.complianceDetails = node.name
-        ? ''
-        : 'Button does not have an accessible name.';
-      break;
-
-    case 'heading':
-      if (!compliance.h1 && node.level === 1) {
-        compliance.h1 = node;
-      }
-      testHeadersAndUpdateCompliance(node);
-      break;
-
-    default:
-      break;
-  }
-}
-
-// checks for common skip link text patters
-const skipLinkRegex =
-  /^(skip(\s|-)?link|main(\s|-)?content|primary(\s|-)?content|page(\s|-)?content|body(\s|-)?content|wrapper|container|app(\s|-)?content|site(\s|-)?content)$/i;
-// check for common non contextual link text
-const nonContextualLinksRegex =
-  /\b(click|click here|here|details|info|more|more info|read more|learn more|go here|this link|check this|tap here|see here|go to link|link)\b[.,!?]*/i;
-
-// to check if it meet the skip link requirement by using the regular expression skipLinkRegex
-function testForSkipLink(node: AccessibilityNode) {
-  const text = node.name?.trim().toLocaleLowerCase();
-  let isSkipLink = false;
-  if (text) {
-    isSkipLink = skipLinkRegex.test(text);
-  }
-  return isSkipLink;
-}
-
-// using the regular expression nonContextualLinksRegex to test if a link has contextual meaning or not
-function testLink(node: AccessibilityNode) {
-  const text = node.name?.trim().toLocaleLowerCase();
-
-  if (!text) {
-    return false;
-  } // Early exit if text is empty or undefined
-
-  const cleanedText = text.replace(/[.,!?]/g, ''); // Remove punctuation
-  return nonContextualLinksRegex.test(cleanedText); // Test against regex
-}
-
-let lastHeadingLevel = 0;
-let firstH1 = false;
-
-// Check the compliance details of each node by validating through the main property/ hierarchy (name and level) of each node
-function testHeadersAndUpdateCompliance(node: AccessibilityNode) {
-  if (!node.name) {
-    node.compliance = false;
-    node.complianceDetails = 'Heading is missing an accessible name.';
-    return;
-  }
-  // Validate heading hierarchy
-  if (node.level !== undefined) {
-    if (node.level === 1) {
-      if (!firstH1) {
-        firstH1 = true;
-      } else {
-        node.compliance = false;
-        node.complianceDetails = 'Only one h1 per page';
-      }
-      lastHeadingLevel = node.level;
-    } else if (lastHeadingLevel !== null && node.level > lastHeadingLevel + 1) {
-      node.compliance = false;
-      node.complianceDetails = `Improper heading hierarchy: ${
-        node.level > lastHeadingLevel
-          ? `Skipped from h${lastHeadingLevel} to h${node.level}.`
-          : `Heading level decreased unexpectedly to h${node.level}.`
-      }`;
-    } else {
-      // Update the last heading level if hierarchy is correct
-      lastHeadingLevel = node.level;
-    }
-  } else {
-    node.compliance = false;
-    node.complianceDetails = 'Heading level is missing.';
-  }
-}
-
-// The SerializedAXNode type represents nodes in an accessibility tree, and its role property defines the role of an element. These roles are based on the ARIA (Accessible Rich Internet Applications) roles and include native HTML roles as well as additional accessibility roles.
-
-// Categories of Roles in ARIA
-// Roles fall into the following categories:
-
-// Abstract Roles: Used only in the ARIA specification and not directly implemented.
-// Widget Roles: Interactive elements like buttons, sliders, and checkboxes.
-// Document Structure Roles: Roles for organizing content, such as headings and sections.
-// Landmark Roles: For high-level structure, like navigation or main content areas.
-// Window Roles: Roles related to the application window, such as dialogs.
-// Common Role Types
-// Here are some common role values you might encounter in the SerializedAXNode:
-
-// --Widget Roles--
-// button
-// checkbox
-// radio
-// slider
-// spinbutton
-// switch
-// textbox
-// combobox
-// listbox
-// menu
-// menuitem
-// menuitemcheckbox
-// menuitemradio
-// progressbar
-// scrollbar
-// tab
-// tabpanel
-// tree
-// treeitem
-// --Document Structure Roles--
-// heading
-// paragraph
-// list
-// listitem
-// table
-// row
-// cell
-// columnheader
-// rowheader
-// grid
-// gridcell
-// article
-// section
-// blockquote
-// --Landmark Roles--
-// banner
-// complementary
-// contentinfo
-// form
-// main
-// navigation
-// region
-// search
-// --Abstract Roles (Not Applicable in Trees)--
-// command
-// composite
-// input
-// landmark
-// range
-// roletype
-// section
-// structure
-// widget
-// --Native HTML Semantics (Automatically Inferred Roles)--
-// link
-// image
-// staticText (often div or span without explicit semantics)
-// generic (no semantic role)
